Refactor and remove redundant code
Refactor and remove redundant code for GPIO setup and button press event handling.

* **Add `gpio_utils.py`**:
  - Create a new module to handle GPIO setup and button press events.
  - Define functions for setting up GPIO pins, handling button press events with debounce logic, and cleaning up GPIO pins.

* **Modify `checkIfDaemonRunning.py`**:
  - Import the new `gpio_utils` module.
  - Replace GPIO setup and button press event handling with functions from `gpio_utils`.
  - Remove redundant code for setting up GPIO pins and handling button press events.

* **Modify `sendData2.py`**:
  - Import the new `gpio_utils` module.
  - Replace GPIO setup and button press event handling with functions from `gpio_utils`.
  - Remove redundant code for setting up GPIO pins and handling button press events.

* **Modify `stateDependant.py`**:
  - Import the new `gpio_utils` module.
  - Replace GPIO setup and button press event handling with functions from `gpio_utils`.
  - Remove redundant code for setting up GPIO pins and handling button press events.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Zeegar/PaiQ?shareId=XXXX-XXXX-XXXX-XXXX).
